Move JS→Java calls to jsToJavaBridgeUtil.ts
